Slightly increase contrast of various elements
I feel this is such neverending story. Every time I stare at this
website for prolonged period of time I find the contrast too low
for comfortable reading so I increase it bit by bit.

What would really help is getting rid of font-weight: 300 for the
whole content but this makes the theme loose its only appeal of
being so slick ...so let's keep postponing that for the next few
years until the looks of sobak.pl will be completely non-fitting
for the current decade :D Then we'll change the theme and ensure
proper contrast.