Fix silly but also the most annoying bug in ages, ugh...
It feels like it's 2012 again and I added whitespace where I shouldn't.
The only difference now is that instead of getting a crash, the
Symfony response class figured that if headers are already sent, it
doens't have to send any headers set within the application itself.

The result of that was that on local all pages were responding with HTTP
200, even the exception page. Redirects still worked because Symfony
prints an HTML code with <meta http-equiv="refresh"> as a fallback.

Another fun aspect is that it works flawlessly on production. Why? My
guess is that on production configs are cached into a single serialized
string and CLI artisan command for "php artisan config:cache" has no
issues with extra whitespace in the config files, it only matters for
HTTP requests. But once the HTTP request hits production, config is
read from unserializing the cache.

One extra line = almost one week of on-and-off debugging after hours :D

I guess some types of errors became so rare these days that I don't
consider them as actually possible anymore. Or at least not without
some log entries.